Fixes #59: Migration to Prisma

+++ src/app/api/cancel-invitation/route.ts
@@ -1,4 +1,5 @@
 import { connectDB } from "@/lib/db";
+import prisma from "@/lib/prisma";
 import { pusherServer } from "@/lib/pusher";
 import { InvitationModel } from "@/models/invitation.model";
 import { InvitationRequest } from "@/types/InvitationRequest";
@@ -18,6 +19,33 @@ export async function POST(req: CustomRequest) {
         { status: 400 }
       );
     }
+    const senderUser = await prisma.user.findUnique({ where: { googleId: sender } });
+    const recipientUser = await prisma.user.findUnique({ where: { googleId: recipient } });
+
+    if (!senderUser || !recipientUser) {
+      return Response.json(new ApiError(400, "Invalid ids"), {
+        status: 400,
+      });
+    }
+
+    const invitation = await prisma.invitation.findFirst({
+      where: {
+        recipientId: recipientUser.id,
+        senderId: recipientUser.id,
+      }
+    })
+
+    if (!invitation) {
+      return Response.json(new ApiError(500, "Error cancelling invitation"), {
+        status: 500,
+      });
+    }
+
+    await prisma.invitation.delete({
+      where: {
+        id: invitation.id,
+      }
+    })
 
     const deletedInvitation = await InvitationModel.findOneAndDelete({
       sender,
